+++ CHANGELOG.md
@@ -6,6 +6,9 @@ _The format is based on [Keep a Changelog](http://keepachangelog.com/) and this 
 
 ## Unreleased
 
+- Fix `theme` prop for styled native components, also fixes `theme` in
+  `defaultProps` for them.
+
 ## [v4.2.1] - 2019-05-30
 
 - Remove className usage checker dev utility due to excessive false-positive noise in certain runtime environments like next.js and the related warning suppression prop (see [#2563](https://github.com/styled-components/styled-components/issues/2563)).